+++ CHANGELOG.md
@@ -25,6 +25,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   using notation like `Mesh[2, 3]` for a 2D manifold in 3D space.
 - Adds adjacency caching to PhysicsNeMo-Mesh `Mesh` objects, allowing efficient
   reuse of neighbor information.
+- Added support for 1D, 2D, and 3D neighborhood attention (natten) via
+  `physicsnemo.nn.functinonal` interface, with full `ShardTensor` support.
 
 ### Changed